requirements.txt
aiohttp>=3.7.4,<4.0.0
discord.py>=2.3.2
Flask>=3.0.0
gunicorn>=21.2.0
line-bot-sdk>=3.5.1
google-generativeai>=0.3.2
python-dotenv>=1.0.0
requests>=2.31.0
psutil>=5.9.0